refactor(Leaderboard): args context -> jotai atom
