+++ app/src/Leaderboard/atoms/leaderboardArgsAtom.tsx
@@ -1,0 +1,5 @@
+import { atom } from 'jotai';
+
+import type { LeaderboardArgs } from '../types/LeaderboardArgs';
+
+export const leaderboardArgsAtom = atom<LeaderboardArgs | null>(null);
