+++ app/src/Leaderboard/types/LeaderboardArgs.d.ts
@@ -1,0 +1,5 @@
+export type LeaderboardArgs = {
+  dateTemplate: DateTemplate;
+  pageNumber: number;
+  promo: number | null;
+};
